requirements.txt
anyio==3.7.1
asyncio==3.4.3
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
platformdirs==3.9.1
python-telegram-bot==20.4
pytz==2023.3
requests==2.31.0
sniffio==1.3.0
soupsieve==2.4.1
tomli==2.0.1
urllib3==2.0.3
yapf==0.40.1
zipp==3.16.2
